bot/requirements.txt
python-telegram-bot
pydub
sqlalchemy
psycopg2
requests
asyncio
aiofiles